refactor(obs): extract the appearance config codec out of the settings view
compressConfig/decompressConfig and the saved-custom-theme reader were defined
in the settings subview, so the OBS URL builder and the overlay parser had to
import a React component to reach them -- the exact edge obsUrl.ts documents
avoiding, and one that drags the Dexie graph into pure codec tests.

Move the codec verbatim to utils/appearanceCodec.ts and repoint every importer.
No re-export is left behind: the five call sites are all updated, so the
settings view no longer pretends to own the format.
